+++ CHANGELOG.md
@@ -4,4 +4,5 @@ All notable changes to this project will be documented in this file.
 
 ## v0.1.0
 
-- 
+- CKEditor Template widget plugin. Drupal CKEditor plugin. Convert multiple
+columns templates into widget.